# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clarift API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentListItem"
                  },
                  "type": "array",
                  "title": "Response List Documents Api V1 Documents Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "description": "Upload a study document (PDF, image, text) for processing.\n\nCreates a Document record, a Job record, and dispatches an ARQ job.\nReturns IDs for polling.",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/practice": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Create Practice",
        "description": "Request targeted practice drills for weak topics.\n\nCreates a PracticeSession record and returns generated drills.",
        "operationId": "create_practice_api_v1_practice_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePracticeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePracticeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/practice/weak-areas": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Get Weak Areas",
        "operationId": "get_weak_areas_api_v1_practice_weak_areas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/practice/{practice_id}": {
      "get": {
        "tags": [
          "practice"
        ],
        "summary": "Get Practice Session",
        "operationId": "get_practice_session_api_v1_practice__practice_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "practice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Practice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
...
      "CreatePracticeRequest": {
        "properties": {
          "weak_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Topics"
          },
          "drill_count": {
            "type": "integer",
            "title": "Drill Count",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "weak_topics"
        ],
        "title": "CreatePracticeRequest",
        "description": "Request body for creating a practice session."
      },
      "CreatePracticeResponse": {
        "properties": {
          "practice_id": {
            "type": "string",
            "title": "Practice Id"
          },
          "drills": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Drills"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "practice_id",
          "drills",
          "message"
        ],
        "title": "CreatePracticeResponse",
        "description": "Response after creating a practice session."
      },
...
      "DocumentListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "created_at"
        ],
        "title": "DocumentListItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PracticeDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "weak_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Topics"
          },
          "drills": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Drills"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "weak_topics",
          "drills",
          "created_at"
        ],
        "title": "PracticeDetailResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}